requirements.txt
aniso8601==9.0.1
astroid==2.5.2
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
curl==0.0.1
fake-useragent==0.1.11
Flask==1.1.2
Flask-JWT-Extended==4.1.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
idna==2.10
importlib-metadata==3.10.0
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.6.0
lxml==4.6.3
MarkupSafe==1.1.1
mccabe==0.6.1
PyJWT==2.0.1
pylint==2.7.4
python-dotenv==0.17.0
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
SQLAlchemy==1.4.5
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.25.8
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.1
